pingo/requirements.txt
aioredis==1.3.1
amqp==2.6.1
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
blessings==1.7
bpython==0.21
Brotli==1.0.9
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.4
celery==4.4.7
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.4
chromedriver-binary==92.0.4515.107.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect2==0.4.1
curtsies==0.3.5
cwcwidth==0.1.4
daphne==3.0.2
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.2
dj-rest-auth==2.1.11
Django==3.2.5
django-allauth==0.45.0
django-appconf==1.0.5
django-axes==5.24.0
django-celery-beat==2.0.0
django-celery-email==3.0.0
django-celery-results==1.2.1
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-discover-runner==1.0
django-easy-pdf==0.1.1
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.2
django-extensions==3.1.3
django-filter==21.1
django-imagekit==4.0.2
django-ipware==4.0.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-mptt==0.13.4
django-nine==0.2.4
django-pandas==0.6.4
django-redis==5.0.0
django-rest-action-permissions==2.0.0
django-restql==0.14.0
django-role-permissions==3.1.1
django-seed==0.2.2
django-settings-export==1.2.1
django-templated-mail==1.1.1
django-timezone-field==4.2.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-flex-fields==0.9.3
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
factory-boy==3.2.0
Faker==8.11.0
filelock==3.0.12
flake8==3.9.2
flower==0.9.2
fonttools==4.27.1
future==0.18.2
greenlet==1.1.1
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
humanize==3.11.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==1.7.0
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
ipinfo==3.0.0
ipinfo-django==1.1.0
itypes==1.2.0
Jinja2==3.0.1
jsonfield==3.1.0
jsonpickle==1.5.2
kombu==4.6.11
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
palettable==3.3.0
pandas==1.3.3
pilkit==2.0
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.2.0
pluggy==1.0.0.dev0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
psycopg2-binary==2.8.5
py==1.10.0
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydyf==0.1.1
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==3.0.0b3
PyPDF2==1.26.0
pyPEG2==2.15.2
pyphen==0.11.0
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
pytest-sugar==0.9.4
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.27
qrcode==7.3.1
redis==3.5.3
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
selenium==3.141.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
squareup==14.1.1.20210915
supervisor==4.2.2
termcolor==1.1.0
text-unidecode==1.3
tinycss2==1.1.0
toml==0.10.2
tornado==5.1.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
validate-email-address==1
vine==1.3.0
virtualenv==20.7.2
virtualenv-clone==0.5.6
wcwidth==0.2.5
weasyprint==53.3
webencodings==0.5.1
Werkzeug==2.0.1
xhtml2pdf==0.2.5
zipp==3.5.0
zope.interface==5.4.0
zopfli==0.1.8
